ci: add coding guideline scan workflow
Add initial workflow for scanning the code base using ECLAIR.
This is the initial step used for integration, at the moment we only
build one single application using west and enable all zephyr coding
guidelines covered by the tool.

This workflow will only run on push events only and will generate a
SARIF file with all violations that can be downloaded and evaluated
locally to help triage and fix issues.

Goal it to address some of the rules with large number of violations
first, get to a manageable number of violations and start posting those
into GH dedicated tracking backend.
Later on we will also start running this on pull requests and flag
violations introduced in PRs.

This should over time also obsolete this existing workflow:

.github/workflows/coding_guidelines.yml

Signed-off-by: Anas Nashif <anas.nashif@intel.com>